Initial Check in
Tool to keep two Visual Studio Solution/Project areas syncronized. The
need: we have files on a network share that's archived and the entire
team can access it; this is where the files need to live. However, we
have SSDs on our PCs and compiling/developing is much faster locally. To
keep the two file systems syncronized, we'll track what solutions are
being kept locally on the PC and monitor the File System with the
FileSystemWatcher class; when something changes in our location, copy up
the change to the network share. (Obviously, exclude \bin\ and \obj\
dirs, since there's no need to keep anything but source files in
parity).

This package uses Compact SQL Server, so the DLLs are deployed with the
VSIX file; it also uses Object List View since it's so easy to use. The
SDF file is included as well, an empty tracking table ships with the
package.

NOTE: Not using the Silverlight tool window as I don't have time to
learn that as well. I've replaced the normal control window with a
standard WinForm since that's easier for me to use for now. At some
point, that form should be converted over to a proper VS Tool Window.
Written in VS2012, but the package should install on VS2010 as well.
It's .NET Framework 4.5 since that's where the IDE Event tracking was
(but now that we're not doing full IDE Event tracking, perhaps it would
run in 4.0 as well?)